pyproject.toml
[project]
name = "resume"
version = "0.1.1"
description = "Resume generator using Typst and Jinja templates"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
    "Jinja2>=3.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["data*", "archive*", "templates*"]
